+++ NEWS.md
@@ -1,6 +1,7 @@
 ## explore 1.1
 
 * abtest() now supports numeric target (t-test)
+* add explain_forest() for binary target
 
 ## explore 1.0.2 (2023-01-14)